mart_api/auth_service/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Muhammad UZair <uzair330@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
uvicorn = "^0.29.0"
fastapi = "^0.111.0"
sqlmodel = "^0.0.16"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
psycopg = {extras = ["binary"], version = "^3.1.18"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
aiokafka = "^0.10.0"
bcrypt = "^4.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"